Build the script with webpack
