+++ docs/tabs.js
@@ -1,240 +1,1 @@
-/**
- * Configure the tabs behavior.
- */
-const jekyllTabsConfiguration = {
-    syncTabsWithSameLabels: false,
-    activateTabFromUrl: false,
-    addCopyToClipboardButton: false,
-    copyToClipboardButtonHtml: '<button>Copy</button>',
-};
-
-const jekyllTabsModule = (function() {
-
-    /**
-     * Remove all "active" classes on li elements that belong to the given ul element.
-     */
-    const removeActiveClasses = function (ulElement) {
-        const liElements = ulElement.querySelectorAll('ul > li');
-
-        Array.prototype.forEach.call(liElements, function(liElement) {
-            liElement.classList.remove('active');
-        });
-    }
-
-    /**
-     * Get the element position looking from the perspective of the parent element.
-     *
-     * Considering the following html:
-     *
-     * <ul>
-     *   <li class="zero">0</li>
-     *   <li class="one">1</li>
-     *   <li class="two">2</li>
-     * </ul>
-     *
-     * Then getChildPosition(document.querySelector('.one')) would return 1.
-     */
-    const getChildPosition = function (element) {
-        const parent = element.parentNode;
-
-        for (let i = 0; i < parent.children.length; i++) {
-            if (parent.children[i] === element) {
-                return i;
-            }
-        }
-
-        throw new Error('No parent found');
-    }
-
-    /**
-     * Returns a list of elements of the given tag that contains the given text.
-     */
-    const findElementsContaining = function(elementTag, text) {
-        const elements = document.querySelectorAll(elementTag);
-        const elementsThatContainText = [];
-
-        for (let i = 0; i < elements.length; i++) {
-            const element = elements[i];
-
-            if (element.textContent.includes(text)) {
-                elementsThatContainText.push(element);
-            }
-        }
-
-        return elementsThatContainText;
-    }
-
-    /**
-     * Handle adding or removing active classes on tab list items.
-     */
-    const handleTabClicked = function(link) {
-        const liTab = link.parentNode;
-        const ulTab = liTab.parentNode;
-        const liPositionInUl = getChildPosition(liTab);
-
-        if (liTab.className.includes('active')) {
-            return;
-        }
-
-        const tabContentId = ulTab.getAttribute('data-tab');
-        const tabContentElement = document.getElementById(tabContentId);
-
-        // Remove all "active" classes first.
-        removeActiveClasses(ulTab);
-        removeActiveClasses(tabContentElement);
-
-        // Then add back active classes depending on the tab (ul element) that was clicked on.
-        tabContentElement.querySelectorAll('ul.tab-content > li')[liPositionInUl].classList.add('active');
-        liTab.classList.add('active');
-    }
-
-    /**
-     * Create a javascript element from html markup.
-     */
-    const createElementFromHtml = function(html) {
-        const template = document.createElement('template');
-        template.innerHTML = html.trim();
-
-        return template.content.firstChild;
-    }
-
-    /**
-     * Copy the given text in the clipboard.
-     *
-     * See https://stackoverflow.com/questions/51805395/navigator-clipboard-is-undefined
-     */
-    const copyToClipboard = function(text) {
-        if (navigator.clipboard && window.isSecureContext) {
-            navigator.clipboard.writeText(text);
-        } else {
-            // Use the 'out of viewport hidden text area' trick
-            const textArea = document.createElement("textarea");
-            textArea.value = text;
-
-            // Move textarea out of the viewport so it's not visible
-            textArea.style.position = "absolute";
-            textArea.style.left = "-999999px";
-
-            document.body.prepend(textArea);
-            textArea.select();
-
-            try {
-                document.execCommand('copy');
-            } catch (error) {
-                console.error(error);
-            } finally {
-                textArea.remove();
-            }
-        };
-    }
-
-    /**
-     * Open the tab specified by a combination of url anchor (#) and query param (?active_tab).
-     *
-     * For example, considering url http://your-jekyll-website.com/some-page/?active_tab=tab-2#my_tabs
-     * Then the tabs with name 'my_tabs' would see tab with label 'tab 2' automatically open.
-     */
-    const activateTabFromUrl = function() {
-        const tabsAnchor = window.location.hash.substring(1);
-
-        if (!tabsAnchor) {
-            return;
-        }
-
-        const targetedTabs = document.getElementById(tabsAnchor);
-
-        if (!targetedTabs) {
-            return;
-        }
-
-        const urlParams = new URLSearchParams(window.location.search);
-        const tabIdToActivate = urlParams.get('active_tab');
-
-        if (!tabIdToActivate) {
-            return;
-        }
-
-        const tabLink = targetedTabs.querySelector('li#' + tabIdToActivate + ' > a');
-
-        if (!tabLink) {
-            return;
-        }
-
-        jekyllTabsModule.handleTabClicked(tabLink);
-    }
-
-    /**
-     * Update the url when clicking on a tab. See method activateTabFromUrl above.
-     */
-    const updateUrlWithActiveTab = function(link) {
-        const liTab = link.parentNode;
-        const ulTab = liTab.parentNode;
-
-        const searchParams = new URLSearchParams(window.location.search);
-        searchParams.set('active_tab', liTab.id);
-
-        const updatedUrl = window.location.pathname + '?' + searchParams.toString() + '#' + ulTab.id;
-        history.replaceState(null, '', updatedUrl);
-    };
-
-    return {
-        findElementsContaining,
-        handleTabClicked,
-        createElementFromHtml,
-        copyToClipboard,
-        activateTabFromUrl,
-        updateUrlWithActiveTab,
-    };
-
-})();
-
-window.addEventListener('load', function () {
-    const tabLinks = document.querySelectorAll('ul.tab > li > a');
-
-    Array.prototype.forEach.call(tabLinks, function(link) {
-        link.addEventListener('click', function (event) {
-            event.preventDefault();
-
-            jekyllTabsModule.handleTabClicked(link);
-
-            if (jekyllTabsConfiguration.activateTabFromUrl) {
-                jekyllTabsModule.updateUrlWithActiveTab(link);
-            }
-
-            if (jekyllTabsConfiguration.syncTabsWithSameLabels) {
-                const linksWithSameName = jekyllTabsModule.findElementsContaining('a', link.textContent);
-
-                for(let i = 0; i < linksWithSameName.length; i++) {
-                    if (linksWithSameName[i] !== link) {
-                        jekyllTabsModule.handleTabClicked(linksWithSameName[i]);
-                    }
-                }
-            }
-        }, false);
-    });
-
-    if (jekyllTabsConfiguration.addCopyToClipboardButton) {
-        const preElements = document.querySelectorAll('ul.tab-content > li pre');
-
-        for(let i = 0; i < preElements.length; i++) {
-            const preElement = preElements[i];
-            const preParentNode = preElement.parentNode;
-            const button = jekyllTabsModule.createElementFromHtml(jekyllTabsConfiguration.copyToClipboardButtonHtml);
-
-            preParentNode.style.position = 'relative';
-            button.style.position = 'absolute';
-            button.style.top = '0px';
-            button.style.right = '0px';
-
-            preParentNode.appendChild(button);
-
-            button.addEventListener('click', function () {
-                jekyllTabsModule.copyToClipboard(preElement.innerText);
-            });
-        }
-    }
-
-    if (jekyllTabsConfiguration.activateTabFromUrl) {
-        jekyllTabsModule.activateTabFromUrl();
-    }
-});
+!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"function"==typeof define&&define.amd?define([],e):"object"==typeof exports?exports.jekyllTabs=e():t.jekyllTabs=e()}(self,(()=>(()=>{var t={918:t=>{t.exports.BY=t=>{const e=t.parentNode;for(let o=0;o<e.children.length;o++)if(e.children[o]===t)return o;throw new Error("No parent found")},t.exports.bX=(t,e)=>{const o=document.querySelectorAll(t),n=[];for(let t=0;t<o.length;t++){const r=o[t];r.textContent.includes(e)&&n.push(r)}return n},t.exports.j4=t=>{const e=document.createElement("template");return e.innerHTML=t.trim(),e.content.firstChild}}},e={};function o(n){var r=e[n];if(void 0!==r)return r.exports;var a=e[n]={exports:{}};return t[n](a,a.exports,o),a.exports}o.d=(t,e)=>{for(var n in e)o.o(e,n)&&!o.o(t,n)&&Object.defineProperty(t,n,{enumerable:!0,get:e[n]})},o.o=(t,e)=>Object.prototype.hasOwnProperty.call(t,e),o.r=t=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})};var n={};return(()=>{"use strict";o.r(n),o.d(n,{init:()=>s});var t=o(918);const e={syncTabsWithSameLabels:!1,activateTabFromUrl:!1,addCopyToClipboardButton:!1,copyToClipboardButtonHtml:"<button>Copy</button>"},r=function(t){const e=t.querySelectorAll("ul > li");Array.prototype.forEach.call(e,(function(t){t.classList.remove("active")}))},a=function(e){const o=e.parentNode,n=o.parentNode,a=(0,t.BY)(o);if(o.className.includes("active"))return;const c=n.getAttribute("data-tab"),l=document.getElementById(c);r(n),r(l),l.querySelectorAll("ul.tab-content > li")[a].classList.add("active"),o.classList.add("active")},c=function(t){if(navigator.clipboard&&window.isSecureContext)navigator.clipboard.writeText(t);else{const e=document.createElement("textarea");e.value=t,e.style.position="absolute",e.style.left="-999999px",document.body.prepend(e),e.select();try{document.execCommand("copy")}catch(t){console.error(t)}finally{e.remove()}}},l=function(){const t=window.location.hash.substring(1);if(!t)return;const e=document.getElementById(t);if(!e)return;const o=new URLSearchParams(window.location.search).get("active_tab");if(!o)return;const n=e.querySelector("li#"+o+" > a");n&&a(n)},i=function(t){const e=t.parentNode,o=e.parentNode,n=new URLSearchParams(window.location.search);n.set("active_tab",e.id);const r=window.location.pathname+"?"+n.toString()+"#"+o.id;history.replaceState(null,"",r)};function s(){window.addEventListener("load",(function(){const o=document.querySelectorAll("ul.tab > li > a");if(Array.prototype.forEach.call(o,(function(o){o.addEventListener("click",(function(n){if(n.preventDefault(),a(o),e.activateTabFromUrl&&i(o),e.syncTabsWithSameLabels){const e=(0,t.bX)("a",o.textContent);for(let t=0;t<e.length;t++)e[t]!==o&&a(e[t])}}),!1)})),e.addCopyToClipboardButton){const o=document.querySelectorAll("ul.tab-content > li pre");for(let n=0;n<o.length;n++){const r=o[n],a=r.parentNode,l=(0,t.j4)(e.copyToClipboardButtonHtml);a.style.position="relative",l.style.position="absolute",l.style.top="0px",l.style.right="0px",a.appendChild(l),l.addEventListener("click",(function(){c(r.innerText)}))}}e.activateTabFromUrl&&l()}))}})(),n})()));
