+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "jekyll-tabs",
+  "version": "1.0.0",
+  "description": "Jekyll Tabs ===========",
+  "private": true,
+  "directories": {
+    "doc": "docs",
+    "lib": "lib"
+  },
+  "scripts": {
+    "test": "jest --env=jsdom --coverage",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Ovski4/jekyll-tabs.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Ovski4/jekyll-tabs/issues"
+  },
+  "homepage": "https://github.com/Ovski4/jekyll-tabs#readme",
+  "devDependencies": {
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
